// src/modules/wishlist/constants/wishlist.enums.ts
export enum WishlistErrorKeyEnum {
  COMPANY_DOMAIN_IS_REQUIRED = 'Company domain is required',
  FAILED_TO_CHECK_EXISTING_WISHLIST = 'Failed to check existing wishlist',
  COMPANY_NOT_FOUND = 'Company not found',
  FAILED_TO_ADD_ITEM_TO_WISHLIST = 'Failed to add item to wishlist',
  FAILED_TO_CREATE_WISHLIST = 'Failed to create wishlist',
  FAILED_TO_FETCH_WISHLIST_INFORMATION = 'Failed to fetch wishlist information',
  WISHLIST_NOT_FOUND = 'Wishlist not found',
}